create employee dashboard
