--- jobfair-frontend/src/components/form-table/columns-type.jsx
@@ -1,19 +1,0 @@
-
-
-export const defaultColumns = [
-    {
-      title: 'ID',
-      dataIndex: 'id',
-      key: 'id',
-    },
-    {
-      title: 'Số điện thoại',
-      dataIndex: 'phone',
-      key: 'phone',
-    },
-    {
-      title: 'Email',
-      dataIndex: 'email',
-      key: 'email',
-    }
-];
