--- jobfair-frontend/src/pages/EmployeeManagementPage/EmployeeRegisterPage.module.scss
@@ -1,1 +1,0 @@
-//handle style scss
