+++ jobfair-frontend/src/pages/EmployeeRegisterPage/EmployeeRegisterPage.style.jsx
@@ -1,0 +1,30 @@
+export const formItemLayout = {
+labelCol: {
+    xs: {
+    span: 24,
+    },
+    sm: {
+    span: 8,
+    },
+},
+wrapperCol: {
+    xs: {
+    span: 24,
+    },
+    sm: {
+    span: 16,
+    },
+},
+};
+export const tailFormItemLayout = {
+    wrapperCol: {
+        xs: {
+        span: 24,
+        offset: 0,
+        },
+        sm: {
+        span: 16,
+        offset: 8,
+        },
+    },
+};
